openclaw/extensions/trading-engine/python/binance/requirements.txt
# Crypto trading engine — runtime dependencies
# Install: pip install -r binance/requirements.txt

# Core — ccxt>=4.5 required for 2026-01-15 Binance signature percent-encoding
ccxt>=4.5
orjson>=3.9
numpy>=1.24
pandas>=2.0
python-dotenv>=1.0
requests>=2.31

# Market analysis
hmmlearn>=0.3
ta>=0.11.0
scikit-learn>=1.5

# Web framework (dashboard + MCP)
fastapi>=0.115
uvicorn>=0.32
pydantic>=2.9
httpx>=0.27
fastapi-mcp>=0.1

# WebSocket (Binance real-time)
websockets>=9.0,<14

# Claude Agent SDK — >=0.1.36 for ThinkingConfig
claude-agent-sdk>=0.1.36
anthropic>=0.50

# Macro regime (FRED data)
yfinance>=0.2.30

# Scheduler (equity cron — optional)
apscheduler>=3.10

# Testing
pytest>=8.0
